+++ lib/json_dynamic_widget.dart
@@ -4,11 +4,18 @@ export 'src/builders/json_widget_builder.dart';
 export 'src/components/functions/json_widget_internal_functions.dart';
 export 'src/components/json_widget_internal_builders.dart';
 export 'src/components/json_widget_registry.dart';
+export 'src/builders/json_widget_internal_builders.dart';
+//
+export 'src/components/functions/json_widget_internal_functions.dart';
+//
+export 'src/components/json_widget_registry.dart';
+//
 export 'src/components/processors/arg_processor.dart';
 export 'src/components/processors/expression_arg_processor.dart';
 export 'src/components/processors/iterable_arg_processor.dart';
 export 'src/components/processors/map_arg_processor.dart';
 export 'src/components/processors/raw_arg_processor.dart';
+//
 export 'src/components/types.dart';
 export 'src/components/values/values.dart';
 //
@@ -23,5 +30,6 @@ export 'src/models/json_widget_data.dart';
 export 'src/models/json_widget_type.dart';
 export 'src/models/named_function_arg.dart';
 export 'src/models/processed_args.dart';
+export 'src/models/widget_value_changed.dart';
 //
 export 'src/widgets/input_error.dart';