+++ lib/src/builders/json_cupertino_switch_builder.dart
@@ -131,7 +131,10 @@ class JsonCupertinoSwitchBuilder extends JsonWidgetBuilder {
   /// [JsonWidgetRegistry].
   @override
   void remove(JsonWidgetData data) {
-    data.registry.removeValue(data.id);
+    data.registry.removeValue(
+      data.id,
+      originator: data.id,
+    );
 
     super.remove(data);
   }
@@ -167,7 +170,11 @@ class JsonCupertinoSwitchBuilder extends JsonWidgetBuilder {
               );
 
               if (data.id.isNotEmpty == true) {
-                data.registry.setValue('${data.id}.error', error ?? '');
+                data.registry.setValue(
+                  '${data.id}.error',
+                  error ?? '',
+                  originator: data.id,
+                );
               }
 
               return error;
@@ -188,7 +195,11 @@ class JsonCupertinoSwitchBuilder extends JsonWidgetBuilder {
                     state.didChange(value);
 
                     if (data.id.isNotEmpty == true) {
-                      data.registry.setValue(data.id, value);
+                      data.registry.setValue(
+                        data.id,
+                        value,
+                        originator: data.id,
+                      );
                     }
                   },
             thumbColor: thumbColor,