+++ lib/src/builders/json_dropdown_button_form_field_builder.dart
@@ -211,7 +211,10 @@ class JsonDropdownButtonFormFieldBuilder extends JsonWidgetBuilder {
   @override
   void remove(JsonWidgetData data) {
     if (data.id.isNotEmpty == true) {
-      data.registry.removeValue(data.id);
+      data.registry.removeValue(
+        data.id,
+        originator: data.id,
+      );
     }
 
     super.remove(data);
@@ -311,8 +314,16 @@ class _JsonDropdownButtonFormFieldWidgetState
         _selectedItemBuilder = (BuildContext context) {
           final widgets = <Widget>[];
           itemEntries.forEach((key, value) {
-            widget.data.registry.setValue('dropdown_item_display', key);
-            widget.data.registry.setValue('dropdown_item_value', value);
+            widget.data.registry.setValue(
+              'dropdown_item_display',
+              key,
+              originator: null,
+            );
+            widget.data.registry.setValue(
+              'dropdown_item_value',
+              value,
+              originator: null,
+            );
 
             widget.builder.selectedItemBuilder.recreate().build(
                   childBuilder: widget.childBuilder,
@@ -377,6 +388,7 @@ class _JsonDropdownButtonFormFieldWidgetState
                   widget.data.registry.setValue(
                     widget.data.id,
                     value,
+                    originator: widget.data.id,
                   );
                 }
               },
@@ -393,8 +405,11 @@ class _JsonDropdownButtonFormFieldWidgetState
                 );
 
                 if (widget.data.id.isNotEmpty == true) {
-                  widget.data.registry
-                      .setValue('${widget.data.id}.error', error ?? '');
+                  widget.data.registry.setValue(
+                    '${widget.data.id}.error',
+                    error ?? '',
+                    originator: widget.data.id,
+                  );
                 }
 
                 return error;