+++ lib/src/builders/json_switch_builder.dart
@@ -227,7 +227,10 @@ class JsonSwitchBuilder extends JsonWidgetBuilder {
   @override
   void remove(JsonWidgetData data) {
     if (data.id.isNotEmpty == true) {
-      data.registry.removeValue(data.id);
+      data.registry.removeValue(
+        data.id,
+        originator: data.id,
+      );
     }
 
     super.remove(data);
@@ -262,7 +265,11 @@ class JsonSwitchBuilder extends JsonWidgetBuilder {
               );
 
               if (data.id.isNotEmpty == true) {
-                data.registry.setValue('${data.id}.error', error ?? '');
+                data.registry.setValue(
+                  '${data.id}.error',
+                  error ?? '',
+                  originator: '${data.id}.error',
+                );
               }
 
               return error;
@@ -296,7 +303,11 @@ class JsonSwitchBuilder extends JsonWidgetBuilder {
                     state.didChange(value);
 
                     if (data.id.isNotEmpty == true) {
-                      data.registry.setValue(data.id, value);
+                      data.registry.setValue(
+                        data.id,
+                        value,
+                        originator: data.id,
+                      );
                     }
                   },
             onInactiveThumbImageError: onInactiveThumbImageError,