+++ lib/src/components/functions/set_value.dart
@@ -16,5 +16,6 @@ class SetValueFunction {
       () => registry.setValue(
             args![0],
             args[1],
+            originator: null,
           );
 }